Struct parser_database::ast::Enum

pub struct Enum {
    pub name: Identifier,
    pub values: Vec<EnumValue, Global>,
    pub attributes: Vec<Attribute, Global>,
    pub span: Span,
    pub inner_span: Span,
    /* private fields */
}

An enum declaration. Enumeration can either be in the database schema, or completely a Prisma level concept.

PostgreSQL stores enums in a schema, while in MySQL the information is in the table definition. On MongoDB the enumerations are handled in the Query Engine.

Fields
name: Identifier

The name of the enum.

enum Foo { ... }
     ^^^

values: Vec<EnumValue, Global>

The values of the enum.

enum Foo {
  Value1
  ^^^^^^
  Value2
  ^^^^^^
}

attributes: Vec<Attribute, Global>

The attributes of this enum.

enum Foo {
  Value1
  Value2

  @@map("1Foo")
  ^^^^^^^^^^^^^
}

span: Span

The location of this enum in the text representation.

inner_span: Span

The span of the inner contents.

Implementations

impl Enum

pub fn iter_values(&self) -> impl ExactSizeIterator