Enum parser_database::OperatorClass

pub enum OperatorClass {

    InetOps,
    JsonbOps,
    JsonbPathOps,
    ArrayOps,
    TextOps,
    BitMinMaxOps,
    VarBitMinMaxOps,
    BpcharBloomOps,
    BpcharMinMaxOps,
    ByteaBloomOps,
    ByteaMinMaxOps,
    DateBloomOps,
    DateMinMaxOps,
    DateMinMaxMultiOps,
    Float4BloomOps,
    Float4MinMaxOps,
    Float4MinMaxMultiOps,
    Float8BloomOps,
    Float8MinMaxOps,
    Float8MinMaxMultiOps,
    InetInclusionOps,
    InetBloomOps,
    InetMinMaxOps,
    InetMinMaxMultiOps,
    Int2BloomOps,
    Int2MinMaxOps,
    Int2MinMaxMultiOps,
    Int4BloomOps,
    Int4MinMaxOps,
    Int4MinMaxMultiOps,
    Int8BloomOps,
    Int8MinMaxOps,
    Int8MinMaxMultiOps,
    NumericBloomOps,
    NumericMinMaxOps,
    NumericMinMaxMultiOps,
    OidBloomOps,
    OidMinMaxOps,
    OidMinMaxMultiOps,
    TextBloomOps,
    TextMinMaxOps,
    TimestampBloomOps,
    TimestampMinMaxOps,
    TimestampMinMaxMultiOps,
    TimestampTzBloomOps,
    TimestampTzMinMaxOps,
    TimestampTzMinMaxMultiOps,
    TimeBloomOps,
    TimeMinMaxOps,
    TimeMinMaxMultiOps,
    TimeTzBloomOps,
    TimeTzMinMaxOps,
    TimeTzMinMaxMultiOps,
    UuidBloomOps,
    UuidMinMaxOps,
    UuidMinMaxMultiOps,

}

Defines operators captured by the index. Used with PostgreSQL GiST/SP-GiST/GIN/BRIN indices.

Variants

InetOps

An operator class for Gist index and inet type.

Indexable Operators

<< (inet, inet)

<<= (inet, inet)

>> (inet, inet)

>>= (inet, inet)

= (inet, inet)

<> (inet, inet)

< (inet, inet)

<= (inet, inet)

> (inet, inet)

>= (inet, inet)

&& (inet, inet)

JsonbOps

An operator class for Gin index and jsonb type.

Indexable Operators

@> (jsonb,jsonb)

@? (jsonb,jsonpath)

@@ (jsonb,jsonpath)

? (jsonb,text)

?| (jsonb,text[])

?& (jsonb,text[])

JsonbPathOps

An operator class for Gin index and jsonb type.

Indexable Operators

@> (jsonb,jsonb)

@? (jsonb,jsonpath)

@@ (jsonb,jsonpath)

ArrayOps

An operator class for Gin index and any array type.

Indexable Operators

&& (anyarray,anyarray)

@> (anyarray,anyarray)

<@ (anyarray,anyarray)

= (anyarray,anyarray)

TextOps

An operator class for SpGist index and text, char and varchar types.

Indexable Operators

= (text,text)

< (text,text)

<= (text,text)

> (text,text)

>= (text,text)

~<~ (text,text)

~<=~ (text,text)

~>=~ (text,text)

~>~ (text,text)

^@ (text,text)

BitMinMaxOps

An operator class for Brin index and bit type.

Indexable Operators

= (bit,bit)

< (bit,bit)

> (bit,bit)

<= (bit,bit)

>= (bit,bit)

VarBitMinMaxOps

An operator class for Brin index and varbit type.

Indexable Operators

= (varbit,varbit)

< (varbit,varbit)

> (varbit,varbit)

<= (varbit,varbit)

>= (varbit,varbit)

BpcharBloomOps

An operator class for Brin index and char type.

Indexable Operators

= (character,character)

BpcharMinMaxOps

An operator class for Brin index and char type.

Indexable Operators

= (character,character)

< (character,character)

<= (character,character)

> (character,character)

>= (character,character)

ByteaBloomOps

An operator class for Brin index and bytea type.

Indexable Operators

= (bytea,bytea)

ByteaMinMaxOps

An operator class for Brin index and bytea type.

Indexable Operators

= (bytea,bytea)

< (bytea,bytea)

<= (bytea,bytea)

> (bytea,bytea)

>= (bytea,bytea)

DateBloomOps

An operator class for Brin index and date type.

Indexable Operators

= (date,date)

DateMinMaxOps

An operator class for Brin index and date type.

Indexable Operators

= (date,date)

< (date,date)

<= (date,date)

> (date,date)

>= (date,date)

DateMinMaxMultiOps

An operator class for Brin index and date type.

Indexable Operators

= (date,date)

< (date,date)

<= (date,date)

> (date,date)

>= (date,date)

Float4BloomOps

An operator class for Brin index and real type.

Indexable Operators

= (float4,float4)

Float4MinMaxOps

An operator class for Brin index and real type.

Indexable Operators

= (float4,float4)

< (float4,float4)

> (float4,float4)

<= (float4,float4)

>= (float4,float4)

Float4MinMaxMultiOps

An operator class for Brin index and real type.

Indexable Operators

= (float4,float4)

< (float4,float4)

> (float4,float4)

<= (float4,float4)

>= (float4,float4)

Float8BloomOps

An operator class for Brin index and double precision type.

Indexable Operators

= (float8,float8)

Float8MinMaxOps

An operator class for Brin index and double precision type.

Indexable Operators

= (float8,float8)

< (float8,float8)

> (float8,float8)

<= (float8,float8)

>= (float8,float8)

Float8MinMaxMultiOps

An operator class for Brin index and double precision type.

Indexable Operators

= (float8,float8)

< (float8,float8)

> (float8,float8)

<= (float8,float8)

>= (float8,float8)

InetInclusionOps

An operator class for Brin index and inet type.

Indexable Operators

<< (inet,inet)

<<= (inet,inet)

>> (inet,inet)

>>= (inet,inet)

= (inet,inet)

&& (inet,inet)

InetBloomOps

An operator class for Brin index and inet type.

Indexable Operators

= (inet,inet)

InetMinMaxOps

An operator class for Brin index and inet type.

Indexable Operators

= (inet,inet)

< (inet,inet)

> (inet,inet)

<= (inet,inet)

>= (inet,inet)

InetMinMaxMultiOps

An operator class for Brin index and inet type.

Indexable Operators

= (inet,inet)

< (inet,inet)

> (inet,inet)

<= (inet,inet)

>= (inet,inet)

Int2BloomOps

An operator class for Brin index and int2 type.

Indexable Operators

= (int2,int2)

Int2MinMaxOps

An operator class for Brin index and int2 type.

Indexable Operators

= (int2,int2)

< (int2,int2)

> (int2,int2)

<= (int2,int2)

>= (int2,int2)

Int2MinMaxMultiOps

An operator class for Brin index and int2 type.

Indexable Operators

= (int2,int2)

< (int2,int2)

> (int2,int2)

<= (int2,int2)

>= (int2,int2)

Int4BloomOps

An operator class for Brin index and int4 type.

Indexable Operators

= (int4,int4)

Int4MinMaxOps

An operator class for Brin index and int4 type.

Indexable Operators

= (int4,int4)

< (int4,int4)

> (int4,int4)

<= (int4,int4)

>= (int4,int4)

Int4MinMaxMultiOps

An operator class for Brin index and int4 type.

Indexable Operators

= (int4,int4)

< (int4,int4)

> (int4,int4)

<= (int4,int4)

>= (int4,int4)

Int8BloomOps

An operator class for Brin index and int8 type.

Indexable Operators

= (int8,int8)

Int8MinMaxOps

An operator class for Brin index and int8 type.

Indexable Operators

= (int8,int8)

< (int8,int8)

> (int8,int8)

<= (int8,int8)

>= (int8,int8)

Int8MinMaxMultiOps

An operator class for Brin index and int8 type.

Indexable Operators

= (int8,int8)

< (int8,int8)

> (int8,int8)

<= (int8,int8)

>= (int8,int8)

NumericBloomOps

An operator class for Brin index and numeric type.

Indexable Operators

= (numeric,numeric)

NumericMinMaxOps

An operator class for Brin index and numeric type.

Indexable Operators

= (numeric,numeric)

< (numeric,numeric)

> (numeric,numeric)

<= (numeric,numeric)

>= (numeric,numeric)

NumericMinMaxMultiOps

An operator class for Brin index and numeric type.

Indexable Operators

= (numeric,numeric)

< (numeric,numeric)

> (numeric,numeric)

<= (numeric,numeric)

>= (numeric,numeric)

OidBloomOps

An operator class for Brin index and oid type.

Indexable Operators

= (oid,oid)

OidMinMaxOps

An operator class for Brin index and oid type.

Indexable Operators

= (oid,oid)

< (oid,oid)

> (oid,oid)

<= (oid,oid)

>= (oid,oid)

OidMinMaxMultiOps

An operator class for Brin index and oid type.

Indexable Operators

= (oid,oid)

< (oid,oid)

> (oid,oid)

<= (oid,oid)

>= (oid,oid)

TextBloomOps

An operator class for Brin index and text type.

Indexable Operators

= (text,text)

TextMinMaxOps

An operator class for Brin index and text type.

Indexable Operators

= (text,text)

< (text,text)

> (text,text)

<= (text,text)

>= (text,text)

TimestampBloomOps

An operator class for Brin index and timestamp type.

Indexable Operators

= (timestamp,timestamp)

TimestampMinMaxOps

An operator class for Brin index and timestamp type.

Indexable Operators

= (timestamp,timestamp)

< (timestamp,timestamp)

> (timestamp,timestamp)

<= (timestamp,timestamp)

>= (timestamp,timestamp)

TimestampMinMaxMultiOps

An operator class for Brin index and timestamp type.

Indexable Operators

= (timestamp,timestamp)

< (timestamp,timestamp)

> (timestamp,timestamp)

<= (timestamp,timestamp)

>= (timestamp,timestamp)

TimestampTzBloomOps

An operator class for Brin index and timestamptz type.

Indexable Operators

= (timestamptz,timestamptz)

TimestampTzMinMaxOps

An operator class for Brin index and timestamptz type.

Indexable Operators

= (timestamptz,timestamptz)

< (timestamptz,timestamptz)

> (timestamptz,timestamptz)

<= (timestamptz,timestamptz)

>= (timestamptz,timestamptz)

TimestampTzMinMaxMultiOps

An operator class for Brin index and timestamptz type.

Indexable Operators

= (timestamptz,timestamptz)

< (timestamptz,timestamptz)

> (timestamptz,timestamptz)

<= (timestamptz,timestamptz)

>= (timestamptz,timestamptz)

TimeBloomOps

An operator class for Brin index and time type.

Indexable Operators

= (time,time)

TimeMinMaxOps

An operator class for Brin index and time type.

Indexable Operators

= (time,time)

< (time,time)

> (time,time)

<= (time,time)

>= (time,time)

TimeMinMaxMultiOps

An operator class for Brin index and time type.

Indexable Operators

= (time,time)

< (time,time)

> (time,time)

<= (time,time)

>= (time,time)

TimeTzBloomOps

An operator class for Brin index and timetz type.

Indexable Operators

= (timetz,timetz)

TimeTzMinMaxOps

An operator class for Brin index and timetz type.

Indexable Operators

= (timetz,timetz)

< (timetz,timetz)

> (timetz,timetz)

<= (timetz,timetz)

>= (timetz,timetz)

TimeTzMinMaxMultiOps

An operator class for Brin index and timetz type.

Indexable Operators

= (timetz,timetz)

< (timetz,timetz)

> (timetz,timetz)

<= (timetz,timetz)

>= (timetz,timetz)

UuidBloomOps

An operator class for Brin index and uuid type.

Indexable Operators

= (uuid,uuid)

UuidMinMaxOps

An operator class for Brin index and uuid type.

Indexable Operators

= (uuid,uuid)

< (uuid,uuid)

> (uuid,uuid)

<= (uuid,uuid)

>= (uuid,uuid)

UuidMinMaxMultiOps

An operator class for Brin index and uuid type.

Indexable Operators

= (uuid,uuid)

< (uuid,uuid)

> (uuid,uuid)

<= (uuid,uuid)

>= (uuid,uuid)

Implementations

impl OperatorClass

pub fn supports_index_type(self, algo: IndexAlgorithm) -> bool

True if the class supports the given index type.