Struct parser_database::ParserDatabase

pub struct ParserDatabase<'ast> { /* fields omitted */ }

ParserDatabase is a container for a Schema AST, together with information gathered during schema validation. Each validation step enriches the database with information that can be used to work with the schema, without changing the AST. Instantiating with ParserDatabase::new() will perform a number of validations and make sure the schema makes sense, but it cannot fail. In case the schema is invalid, diagnostics will be created and the resolved information will be incomplete.

Validations are carried out in the following order:

The AST is walked a first time to resolve names: to each relevant identifier, we attach an ID that can be used to reference the corresponding item (model, enum, field, …)

The AST is walked a second time to resolve types. For each field and each type alias, we look at the type identifier and resolve what it refers to.

The AST is walked a third time to validate attributes on models and fields.

Global validations are then performed on the mostly validated schema. Currently only index name collisions.

Lifetimes

Throughout the ParserDatabase implementation, you will see many lifetime annotations. The only significant lifetime is the lifetime of the reference to the AST contained in ParserDatabase, that we call by convention 'ast. Apart from that, everything should be owned or locally borrowed, to keep lifetime management simple.

Implementations

impl<'ast> ParserDatabase<'ast>

pub fn walk_model(&self, model_id: ModelId) -> ModelWalker<'ast, '_>

Find a model by ID.

pub fn walk_models(&self) -> impl Iterator<Item = ModelWalker<'ast, '_>> + '_

Walk all the models in the schema.

pub fn walk_composite_type(
&self,
ctid: CompositeTypeId
) -> CompositeTypeWalker<'ast, '_>

Walk a specific composite type by ID.

pub fn walk_composite_types(
&self
) -> impl Iterator<Item = CompositeTypeWalker<'ast, '_>> + '_

Walk all the composite types in the schema.

pub fn walk_relations(
&self
) -> impl Iterator<Item = RelationWalker<'ast, '_>> + '_

Walk all the relations in the schema. A relation may be defined by one or two fields; in both cases, it is still a single relation.

pub fn walk_complete_inline_relations(
&self
) -> impl Iterator<Item = CompleteInlineRelationWalker<'ast, '_>> + '_

Iterate all complete relations that are not many to many and are correctly defined from both sides.

impl<'ast> ParserDatabase<'ast>

pub fn new(
ast: &'ast SchemaAst,
diagnostics: Diagnostics
) -> (Self, Diagnostics)

See the docs on ParserDatabase.

pub fn alias_scalar_field_type(&self, alias_id: &AliasId) -> &ScalarFieldType

pub fn ast(&self) -> &'ast SchemaAst

pub fn find_model_field(
&self,
model_id: ModelId,
field_name: &str
) -> Option<FieldId>

pub fn get_enum_database_name(&self, enum_id: EnumId) -> Option<&'ast str>

pub fn get_enum_value_database_name(
&self,
enum_id: EnumId,
value_idx: u32
) -> Option<&'ast str>