Struct parser_database::walkers::CompositeTypeWalker

pub struct CompositeTypeWalker<'db> { /* fields omitted */ }

A composite type, introduced with the type keyword in the schema.

Example:

type Address {
    name String
    streetName String
    streetNumber Int
    city String
    zipCode Int
    countryCode String
}

Implementations

impl<'db> CompositeTypeWalker<'db>

pub fn composite_type_id(self) -> CompositeTypeId

The ID of the composite type node in the AST.

pub fn ast_composite_type(self) -> &'db CompositeType

The composite type node in the AST.

pub fn name(self) -> &'db str

The name of the composite type in the schema.

pub fn field(self, field_id: FieldId) -> CompositeTypeFieldWalker<'db>

Get the field with the given ID.

pub fn fields(self) -> impl Iterator<Item = CompositeTypeFieldWalker<'db>>

Iterator over all the fields of the composite type.