Struct parser_database::walkers::IndexWalker

pub struct IndexWalker<'db> { /* private fields */ }

An index, unique or fulltext attribute.

Implementations

impl<'db> IndexWalker<'db>

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the index.

@@index([a, b], map: "theName")
                     ^^^^^^^^^

pub fn attribute_name(self) -> &'static str

The attribute name: "index" for @@unique, "fulltext" for @@fultext and "index" for @index and @@index.

pub fn attribute_id(self) -> AttributeId

The id of the index, if explicitly defined.

pub fn index_type(self) -> IndexType

The index type.

pub fn name(self) -> Option<&'db str>

The name argument of the index attribute. The client name.

@@index([a, b], name: "theName")
                     ^^^^^^^^^

pub fn algorithm(self) -> Option<IndexAlgorithm>

The index algorithm, if a specific one was specified for the index.

pub fn ast_attribute(self) -> &'db Attribute

The AST node of the index/unique attribute.

pub fn all_field_names(self) -> impl Iterator<Item = &'db str>

Iterate over all the names in all the paths in the fields argument.

For example, @@index([a, b.c.d]) would return an iterator over “a”, “b”, “c”, “d”.

pub fn fields(self) -> impl ExactSizeIterator<Item = IndexFieldWalker<'db>>

The scalar fields covered by the index.

pub fn scalar_field_attributes(
self
) -> impl ExactSizeIterator<Item = ScalarFieldAttributeWalker<'db>>

The scalar fields covered by the index, and their arguments.

pub fn contains_field(self, field: ScalarFieldWalker<'db>) -> bool

True, if given field is a part of the indexed fields.

pub fn contains_exactly_the_fields(
self,
fields: impl ExactSizeIterator<Item = ScalarFieldAttributeWalker<'db>>
) -> bool

True if the field contains exactly the same fields in the same order, and with the same attributes.

pub fn is_defined_on_field(self) -> bool

Whether the index is defined on a single field (otherwise: on the model).

pub fn is_unique(self) -> bool

Is this an @@unique?

pub fn is_fulltext(self) -> bool

Is this a @@fulltext?

pub fn is_normal(self) -> bool

Is this an @@index?

pub fn clustered(self) -> Option<bool>

If true, the index defines the storage and ordering of the row. Mostly matters on SQL Server where one can change the clustering.

pub fn model(self) -> ModelWalker<'db>

The model the index is defined on.

pub fn source_field(self) -> Option<ScalarFieldWalker<'db>>

The field the model was defined on, if any.