Struct parser_database::walkers::PrimaryKeyWalker

pub struct PrimaryKeyWalker<'db> { /* private fields */ }

An @(@)id attribute in the schema.

Implementations

impl<'db> PrimaryKeyWalker<'db>

pub fn ast_attribute(self) -> &'db Attribute

The @(@)id AST node.

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the id.

@@id([a, b], map: "theName")
                  ^^^^^^^^^

pub fn is_defined_on_field(self) -> bool

Is this an @id on a specific field, rather than on the model?

pub fn attribute_name(self) -> &'static str

If defined on a specific field, returns @id. Otherwise @@id.

pub fn clustered(self) -> Option<bool>

If true, the index defines the storage and ordering of the row. Mostly matters on SQL Server where one can change the clustering.

pub fn model(self) -> ModelWalker<'db>

The model the id is deined on.

pub fn name(self) -> Option<&'db str>

The name argument of the id attribute. The client name.

@@id([a, b], name: "theName")
                   ^^^^^^^^^

pub fn fields(
self
) -> impl ExactSizeIterator<Item = ScalarFieldWalker<'db>> + 'db

The scalar fields constrained by the id.

pub fn scalar_field_attributes(
self
) -> impl ExactSizeIterator<Item = ScalarFieldAttributeWalker<'db>> + 'db

The scalar fields covered by the id, and their arguments.

pub fn contains_exactly_fields(
self,
fields: impl ExactSizeIterator<Item = ScalarFieldWalker<'db>>
) -> bool

Do the constrained fields match exactly these?