Struct parser_database::walkers::RelationFieldWalker

pub struct RelationFieldWalker<'ast, 'db> { /* fields omitted */ }

A relation field on a model in the schema.

Implementations

impl<'ast, 'db> RelationFieldWalker<'ast, 'db>

pub fn field_id(self) -> FieldId

The ID of the AST node of the field.

pub fn foreign_key_name(self) -> Option<&'ast str>

The foreign key name of the relation (@relation(map: ...)).

pub fn name(self) -> &'ast str

The field name.

pub fn ast_field(self) -> &'ast Field

The AST node of the field.

pub fn explicit_on_delete(self) -> Option<ReferentialAction>

The onDelete argument on the relation.

pub fn explicit_on_update(self) -> Option<ReferentialAction>

The onUpdate argument on the relation.

pub fn explicit_relation_name(self) -> Option<&'ast str>

The relation name explicitly written in the schema source.

pub fn is_ignored(self) -> bool

Is there an @ignore attribute on the field?

pub fn model(self) -> ModelWalker<'ast, 'db>

The model containing the field.

pub fn relation_attribute(self) -> Option<&'ast Attribute>

The @relation attribute in the field AST.

pub fn related_model(self) -> ModelWalker<'ast, 'db>

The model referenced by the relation.

pub fn referenced_fields(
self
) -> Option<impl ExactSizeIterator<Item = ScalarFieldWalker<'ast, 'db>> + 'db>

The fields in the @relation(references: ...) argument.

pub fn relation_name(self) -> RelationName<'ast>

The name of the relation. Either uses the name (or default) argument, or generates an implicit name.

pub fn referential_arity(self) -> FieldArity

pub fn references_singular_id_field(self) -> bool

Used for validation.

pub fn referencing_fields(
self
) -> Option<impl ExactSizeIterator<Item = ScalarFieldWalker<'ast, 'db>>>

pub fn fields(
self
) -> Option<impl ExactSizeIterator<Item = ScalarFieldWalker<'ast, 'db>>>