Struct parser_database::walkers::Walker

pub struct Walker<'db, I> {
    pub db: &'db ParserDatabase,
    pub id: I,
}

A generic walker. Only walkers intantiated with a concrete ID type (I) are useful.

Fields
db: &'db ParserDatabase

The parser database being traversed.

id: I

The identifier of the focused element.

Implementations

impl<'db> Walker<'db, CompositeTypeId>

pub fn composite_type_id(self) -> CompositeTypeId

The ID of the composite type node in the AST.

pub fn ast_composite_type(self) -> &'db CompositeType

The composite type node in the AST.

pub fn name(self) -> &'db str

The name of the composite type in the schema.

pub fn fields(
self
) -> impl ExactSizeIterator<Item = CompositeTypeFieldWalker<'db>> + Clone

Iterator over all the fields of the composite type.

impl<'db> Walker<'db, (CompositeTypeId, FieldId)>

pub fn ast_field(self) -> &'db Field

The AST node for the field.

pub fn composite_type(self) -> CompositeTypeWalker<'db>

The composite type containing the field.

pub fn documentation(&self) -> Option<&str>

The optional documentation string of the field.

pub fn mapped_name(self) -> Option<&'db str>

The name contained in the @map() attribute of the field, if any.

pub fn field_id(self) -> FieldId

The ID of the field in the AST.

pub fn name(self) -> &'db str

The name of the field.

pub fn arity(self) -> FieldArity

Is the field required, optional or a list?

pub fn type(self) -> &'db ScalarFieldType

The type of the field, e.g. String in streetName String?.

pub fn scalar_type(self) -> Option<ScalarType>

The type of the field in case it is a scalar type (not an enum, not a composite type).

pub fn default_attribute(self) -> Option<&'db Attribute>

The @default() AST attribute on the field, if any.

pub fn raw_native_type(
self
) -> Option<(&'db str, &'db str, &'db [String], Span)>

(attribute scope, native type name, arguments, span)

For example: @db.Text would translate to (“db”, “Text”, &[], )

pub fn default_value(self) -> Option<&'db Expression>

The value expression in the @default attribute.

score Int @default(0)
                   ^

pub fn default_mapped_name(self) -> Option<&'db str>

The mapped name of the default value. Always None in composite types at the moment.

name String @default("george", map: "name_default_to_george")
                                    ^^^^^^^^^^^^^^^^^^^^^^^^

pub fn database_name(self) -> &'db str

The final database name of the field. See crate docs for explanations on database names.

impl<'db> Walker<'db, EnumId>

pub fn name(self) -> &'db str

The name of the enum.

pub fn ast_enum(self) -> &'db Enum

The AST node.

pub fn database_name(self) -> &'db str

The database name of the enum.

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the enum:

enum Colour {
    RED
    GREEN
    BLUE

    @@map("Color")
          ^^^^^^^
}

pub fn values(self) -> impl ExactSizeIterator<Item = EnumValueWalker<'db>>

The values of the enum.

pub fn indentation(self) -> IndentationType

How fields are indented in the enum.

pub fn newline(self) -> NewlineType

What kind of newlines the enum uses.

pub fn schema(self) -> Option<(&'db str, Span)>

The name of the schema the enum belongs to.

@@schema("public")
         ^^^^^^^^

impl<'db> Walker<'db, (EnumId, usize)>

pub fn documentation(self) -> Option<&'db str>

The enum documentation

pub fn name(self) -> &'db str

The name of the value.

pub fn database_name(self) -> &'db str

The database name of the enum.

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the value:

enum Colour {
    RED @map("scarlet")
    GREEN @map("salad")
               ^^^^^^^
    BLUE @map("schmurf")
}

impl<'db> Walker<'db, (ModelId, FieldId)>

pub fn ast_field(self) -> &'db Field

The AST node for the field.

pub fn name(self) -> &'db str

The field name.

pub fn model(self) -> ModelWalker<'db>

Traverse the field’s parent model.

pub fn refine(self) -> RefinedFieldWalker<'db>

Find out which kind of field this is.

impl<'db> Walker<'db, ModelId>

pub fn name(self) -> &'db str

The name of the model.

pub fn fields(self) -> impl ExactSizeIterator<Item = FieldWalker<'db>> + Clone

Traverse the fields of the models in the order they were defined.

pub fn field_is_indexed_for_autoincrement(self, field_id: FieldId) -> bool

Whether MySQL would consider the field indexed for autoincrement purposes.

pub fn field_is_single_pk(self, field: FieldId) -> bool

Whether the field is the whole primary key. Will match @id and @@id([fieldName]).

pub fn field_is_part_of_a_compound_pk(self, field: FieldId) -> bool

Is the field part of a compound primary key.

pub fn model_id(self) -> ModelId

The ID of the model in the db

pub fn ast_model(self) -> &'db Model

The AST node.

pub fn is_ignored(self) -> bool

Model has the @@ignore attribute.

pub fn database_name(self) -> &'db str

The name of the database table the model points to.

pub fn get_field_database_name(self, field_id: FieldId) -> &'db str

Get the database name of the scalar field.

pub fn get_field_database_names(
self,
fields: &'db [FieldId]
) -> impl Iterator<Item = &'db str>

Get the database names of the constrained scalar fields.

pub fn has_single_id_field(self) -> bool

Used in validation. True only if the model has a single field id.

pub fn mapped_name(self) -> Option<&'db str>

The name in the @@map attribute.

pub fn primary_key(self) -> Option<PrimaryKeyWalker<'db>>

The primary key of the model, if defined.

pub fn scalar_field(self, field_id: FieldId) -> ScalarFieldWalker<'db>

Walk a scalar field by id.

pub fn scalar_fields(self) -> impl Iterator<Item = ScalarFieldWalker<'db>>

Iterate all the scalar fields in a given model in the order they were defined.

pub fn unique_criterias(self) -> impl Iterator<Item = UniqueCriteriaWalker<'db>>

All unique criterias of the model; consisting of the primary key and unique indexes, if set.

pub fn indexes(self) -> impl Iterator<Item = IndexWalker<'db>>

Iterate all the indexes in the model in the order they were defined.

pub fn relation_fields(self) -> impl Iterator<Item = RelationFieldWalker<'db>>

All (concrete) relation fields of the model.

pub fn relation_field(self, field_id: FieldId) -> RelationFieldWalker<'db>

Find a relation field with the given id.

Panics

If the field does not exist.

pub fn relations_from(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that start from this model.

pub fn relations_to(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that reference this model.

pub fn inline_relations_from(
self
) -> impl Iterator<Item = InlineRelationWalker<'db>>

1:n and 1:1 relations that start from this model.

pub fn complete_inline_relations_from(
self
) -> impl Iterator<Item = CompleteInlineRelationWalker<'db>>

1:n and 1:1 relations, starting from this model and having both sides defined.

pub fn indentation(self) -> IndentationType

How fields and arguments are indented in the model.

pub fn newline(self) -> NewlineType

What kind of newlines the model uses.

pub fn schema(self) -> Option<(&'db str, Span)>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^

pub fn schema_name(self) -> Option<&'db str>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^

impl<'db> Walker<'db, ManyToManyRelationId>

pub fn is_self_relation(self) -> bool

Is this a many-to-many self-relation?

pub fn model_a(self) -> ModelWalker<'db>

The model which comes first in the alphabetical order.

pub fn model_b(self) -> ModelWalker<'db>

The model which comes after model a in the alphabetical order.

pub fn field_a(self) -> RelationFieldWalker<'db>

The field that defines the relation in model a.

pub fn field_b(self) -> RelationFieldWalker<'db>

The field that defines the relation in model b.

pub fn as_relation(self) -> RelationWalker<'db>

Traverse this relation as a generic relation.

pub fn relation_name(self) -> RelationName<'db>

The name of the relation.

impl<'db> Walker<'db, RelationId>

pub fn is_self_relation(self) -> bool

Is this a relation where both ends are the same model?

pub fn refine(self) -> RefinedRelationWalker<'db>

Converts the walker to either an implicit many to many, or a inline relation walker gathering 1:1 and 1:n relations.

pub fn explicit_relation_name(self) -> Option<&'db str>

The relation name in the schema.

myField OtherModel @relation("thisModelToOtherModel", fields: [fkfield], references: [id])
//                           ^^^^^^^^^^^^^^^^^^^^^^^

impl<'db, I> Walker<'db, I>

pub fn walk<J>(self, other: J) -> Walker<'db, J>

Traverse something else in the same schema.