Type Alias parser_database::walkers::CompositeTypeWalker

pub type CompositeTypeWalker<'db> = Walker<'db, CompositeTypeId>;

A composite type, introduced with the type keyword in the schema.

Example:

type Address {
    name String
    streetName String
    streetNumber Int
    city String
    zipCode Int
    countryCode String
}

Aliased Type

struct CompositeTypeWalker<'db> {
    pub db: &'db ParserDatabase,
    pub id: CompositeTypeId,
}

Fields
db: &'db ParserDatabase

The parser database being traversed.

id: CompositeTypeId

The identifier of the focused element.

Implementations

impl<'db> CompositeTypeWalker<'db>

pub fn composite_type_id(self) -> CompositeTypeId

The ID of the composite type node in the AST.

pub fn ast_composite_type(self) -> &'db CompositeType

The composite type node in the AST.

pub fn name(self) -> &'db str

The name of the composite type in the schema.

pub fn fields( self ) -> impl ExactSizeIterator<Item = CompositeTypeFieldWalker<'db>> + Clone

Iterator over all the fields of the composite type.

impl<'db, I> Walker<'db, I>

pub fn walk<J>(self, other: J) -> Walker<'db, J>

Traverse something else in the same schema.