Type Definition parser_database::walkers::ImplicitManyToManyRelationWalker

pub type ImplicitManyToManyRelationWalker<'db> = Walker<'db, ManyToManyRelationId>;

Describes an implicit m:n relation between two models. Neither side defines fields, attributes or referential actions, which are all inferred by Prisma.

Implementations

impl<'db> ImplicitManyToManyRelationWalker<'db>

pub fn model_a(self) -> ModelWalker<'db>

The model which comes first in the alphabetical order.

pub fn model_b(self) -> ModelWalker<'db>

The model which comes after model a in the alphabetical order.

pub fn field_a(self) -> RelationFieldWalker<'db>

The field that defines the relation in model a.

pub fn field_b(self) -> RelationFieldWalker<'db>

The field that defines the relation in model b.

pub fn relation_name(self) -> RelationName<'db>

The name of the relation.