Struct parser_database::walkers::Walker

pub struct Walker<'db, I> {
    pub db: &'db ParserDatabase,
    pub id: I,
}

A generic walker. Only walkers intantiated with a concrete ID type (I) are useful.

Fields
db: &'db ParserDatabase

The parser database being traversed.

id: I

The identifier of the focused element.

Implementations

impl<'db> Walker<'db, EnumId>

pub fn name(self) -> &'db str

The name of the enum.

pub fn ast_enum(self) -> &'db Enum

The AST node.

pub fn database_name(self) -> &'db str

The database name of the enum.

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the enum:

enum Colour {
    RED
    GREEN
    BLUE

    @@map("Color")
          ^^^^^^^
}

pub fn values(self) -> impl ExactSizeIterator<Item = EnumValueWalker<'db>>

The values of the enum.

pub fn schema(self) -> Option<(&'db str, Span)>

The name of the schema the enum belongs to.

@@schema("public")
         ^^^^^^^^

impl<'db> Walker<'db, (EnumId, usize)>

pub fn documentation(self) -> Option<&'db str>

The enum documentation

pub fn name(self) -> &'db str

The name of the value.

pub fn database_name(self) -> &'db str

The database name of the enum.

pub fn mapped_name(self) -> Option<&'db str>

The mapped name of the value:

enum Colour {
    RED @map("scarlet")
    GREEN @map("salad")
               ^^^^^^^
    BLUE @map("schmurf")
}

impl<'db> Walker<'db, ModelId>

pub fn name(self) -> &'db str

The name of the model.

pub fn field_is_indexed_for_autoincrement(self, field_id: FieldId) -> bool

Whether MySQL would consider the field indexed for autoincrement purposes.

pub fn field_is_single_pk(self, field: FieldId) -> bool

Whether the field is the whole primary key. Will match @id and @@id([fieldName]).

pub fn field_is_part_of_a_compound_pk(self, field: FieldId) -> bool

Is the field part of a compound primary key.

pub fn model_id(self) -> ModelId

The ID of the model in the db

pub fn ast_model(self) -> &'db Model

The AST node.

pub fn is_ignored(self) -> bool

Model has the @@ignore attribute.

pub fn database_name(self) -> &'db str

The name of the database table the model points to.

pub fn get_field_database_name(self, field_id: FieldId) -> &'db str

Get the database name of the scalar field.

pub fn get_field_database_names(
self,
fields: &'db [FieldId]
) -> impl Iterator<Item = &'db str>

Get the database names of the constrained scalar fields.

pub fn has_single_id_field(self) -> bool

Used in validation. True only if the model has a single field id.

pub fn mapped_name(self) -> Option<&'db str>

The name in the @@map attribute.

pub fn primary_key(self) -> Option<PrimaryKeyWalker<'db>>

The primary key of the model, if defined.

pub fn scalar_field(self, field_id: FieldId) -> ScalarFieldWalker<'db>

Walk a scalar field by id.

pub fn scalar_fields(self) -> impl Iterator<Item = ScalarFieldWalker<'db>>

Iterate all the scalar fields in a given model in the order they were defined.

pub fn unique_criterias(self) -> impl Iterator<Item = UniqueCriteriaWalker<'db>>

All unique criterias of the model; consisting of the primary key and unique indexes, if set.

pub fn indexes(self) -> impl Iterator<Item = IndexWalker<'db>>

Iterate all the indexes in the model in the order they were defined.

pub fn relation_fields(self) -> impl Iterator<Item = RelationFieldWalker<'db>>

All (concrete) relation fields of the model.

pub fn relation_field(self, field_id: FieldId) -> RelationFieldWalker<'db>

Find a relation field with the given id.

Panics

If the field does not exist.

pub fn relations_from(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that start from this model.

pub fn relations_to(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that reference this model.

pub fn inline_relations_from(
self
) -> impl Iterator<Item = InlineRelationWalker<'db>>

1:n and 1:1 relations that start from this model.

pub fn complete_inline_relations_from(
self
) -> impl Iterator<Item = CompleteInlineRelationWalker<'db>>

1:n and 1:1 relations, starting from this model and having both sides defined.

pub fn indentation(self) -> IndentationType

How fields and arguments are indented in the model.

pub fn newline(self) -> NewlineType

What kind of newlines the model uses.

pub fn schema(self) -> Option<(&'db str, Span)>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^

pub fn schema_name(self) -> Option<&'db str>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^

impl<'db> Walker<'db, ManyToManyRelationId>

pub fn model_a(self) -> ModelWalker<'db>

The model which comes first in the alphabetical order.

pub fn model_b(self) -> ModelWalker<'db>

The model which comes after model a in the alphabetical order.

pub fn field_a(self) -> RelationFieldWalker<'db>

The field that defines the relation in model a.

pub fn field_b(self) -> RelationFieldWalker<'db>

The field that defines the relation in model b.

pub fn relation_name(self) -> RelationName<'db>

The name of the relation.

impl<'db> Walker<'db, RelationId>

pub fn refine(self) -> RefinedRelationWalker<'db>

Converts the walker to either an implicit many to many, or a inline relation walker gathering 1:1 and 1:n relations.