Type Definition parser_database::walkers::ModelWalker

pub type ModelWalker<'db> = Walker<'db, ModelId>;

A model declaration in the Prisma schema.

Implementations

impl<'db> ModelWalker<'db>

pub fn name(self) -> &'db str

The name of the model.

pub fn fields(self) -> impl ExactSizeIterator<Item = FieldWalker<'db>> + Clone

Traverse the fields of the models in the order they were defined.

pub fn field_is_indexed_for_autoincrement(self, field_id: FieldId) -> bool

Whether MySQL would consider the field indexed for autoincrement purposes.

pub fn field_is_single_pk(self, field: FieldId) -> bool

Whether the field is the whole primary key. Will match @id and @@id([fieldName]).

pub fn field_is_part_of_a_compound_pk(self, field: FieldId) -> bool

Is the field part of a compound primary key.

pub fn model_id(self) -> ModelId

The ID of the model in the db

pub fn ast_model(self) -> &'db Model

The AST node.

pub fn is_ignored(self) -> bool

Model has the @@ignore attribute.

pub fn database_name(self) -> &'db str

The name of the database table the model points to.

pub fn has_single_id_field(self) -> bool

Used in validation. True only if the model has a single field id.

pub fn mapped_name(self) -> Option<&'db str>

The name in the @@map attribute.

pub fn primary_key(self) -> Option<PrimaryKeyWalker<'db>>

The primary key of the model, if defined.

pub fn scalar_fields(self) -> impl Iterator<Item = ScalarFieldWalker<'db>>

Iterate all the scalar fields in a given model in the order they were defined.

pub fn unique_criterias(self) -> impl Iterator<Item = UniqueCriteriaWalker<'db>>

All unique criterias of the model; consisting of the primary key and unique indexes, if set.

pub fn required_unique_criterias(
self
) -> impl Iterator<Item = UniqueCriteriaWalker<'db>>

All required unique criterias of the model; consisting of the primary key and unique indexes, if set.

pub fn indexes(self) -> impl Iterator<Item = IndexWalker<'db>>

Iterate all the indexes in the model in the order they were defined.

pub fn relation_fields(
self
) -> impl Iterator<Item = RelationFieldWalker<'db>> + Clone + 'db

All (concrete) relation fields of the model.

pub fn relations_from(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that start from this model.

pub fn relations_to(self) -> impl Iterator<Item = RelationWalker<'db>>

All relations that reference this model.

pub fn inline_relations_from(
self
) -> impl Iterator<Item = InlineRelationWalker<'db>>

1:n and 1:1 relations that start from this model.

pub fn complete_inline_relations_from(
self
) -> impl Iterator<Item = CompleteInlineRelationWalker<'db>>

1:n and 1:1 relations, starting from this model and having both sides defined.

pub fn indentation(self) -> IndentationType

How fields and arguments are indented in the model.

pub fn newline(self) -> NewlineType

What kind of newlines the model uses.

pub fn schema(self) -> Option<(&'db str, Span)>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^

pub fn schema_name(self) -> Option<&'db str>

The name of the schema the model belongs to.

@@schema("public")
         ^^^^^^^^